+++ pkg/worker/emulator/libretro/frontend_test.go
@@ -1,4 +1,4 @@
-package nanoarch
+package libretro
 
 import (
 	"fmt"
@@ -230,6 +230,35 @@ func TestStateConcurrency(t *testing.T) {
 }
 
 // lucky returns random boolean.
-func lucky() bool {
-	return rand.Intn(2) == 1
+func lucky() bool { return rand.Intn(2) == 1 }
+
+func TestConcurrentInput(t *testing.T) {
+	players := NewGameSessionInput()
+
+	events := 1000
+	var wg sync.WaitGroup
+
+	wg.Add(events * 2)
+
+	go func() {
+		for i := 0; i < events; i++ {
+			player := rand.Intn(maxPort)
+			go func() {
+				players.setInput(player, []byte{0, 1})
+				wg.Done()
+			}()
+		}
+	}()
+
+	go func() {
+		for i := 0; i < events; i++ {
+			player := rand.Intn(maxPort)
+			go func() {
+				players.isKeyPressed(uint(player), 100)
+				wg.Done()
+			}()
+		}
+	}()
+
+	wg.Wait()
 }
