+++ pkg/worker/emulator/libretro/libretro.h
