+++ pkg/worker/emulator/libretro/nanoarch_test.go
@@ -1,4 +1,4 @@
-package nanoarch
+package libretro
 
 import (
 	"crypto/md5"
