+++ pkg/worker/emulator/libretro/storage.go
@@ -1,4 +1,4 @@
-package nanoarch
+package libretro
 
 import (
 	"os"
